db/requirement.txt
requests
psycopg2-binary
dotenv
pandas